+++ http/service_test.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"io/ioutil"
 	"net/http"
+	"net/http/httptest"
 	"net/url"
 	"os"
 	"strings"
@@ -826,36 +827,90 @@ func Test_RegisterStatus(t *testing.T) {
 }
 
 func Test_FormRedirect(t *testing.T) {
-	m := &MockStore{}
-	c := &mockClusterService{}
+	m := &MockStore{
+		leaderAddr: "foo:4002",
+	}
+	c := &mockClusterService{
+		apiAddr: "http://foo:4001",
+	}
 
 	s := New("127.0.0.1:0", m, c, nil)
 	req := mustNewHTTPRequest("http://qux:4001")
 
-	if rd := s.FormRedirect(req, "http://foo:4001"); rd != "http://foo:4001" {
-		t.Fatal("failed to form redirect for simple URL")
+	rd, err := s.FormRedirect(req)
+	if err != nil {
+		t.Fatalf("failed to form redirect: %s", err.Error())
+	}
+	if exp, got := "http://foo:4001", rd; exp != got {
+		t.Fatalf("incorrect redirect, exp: %s, got: %s", exp, got)
 	}
 }
 
 func Test_FormRedirectParam(t *testing.T) {
-	m := &MockStore{}
-	c := &mockClusterService{}
+	m := &MockStore{
+		leaderAddr: "foo:4002",
+	}
+	c := &mockClusterService{
+		apiAddr: "http://foo:4001",
+	}
 	s := New("127.0.0.1:0", m, c, nil)
 	req := mustNewHTTPRequest("http://qux:4001/db/query?x=y")
 
-	if rd := s.FormRedirect(req, "http://foo:4001"); rd != "http://foo:4001/db/query?x=y" {
-		t.Fatal("failed to form redirect for URL")
+	rd, err := s.FormRedirect(req)
+	if err != nil {
+		t.Fatalf("failed to form redirect: %s", err.Error())
+	}
+
+	if exp, got := "http://foo:4001/db/query?x=y", rd; rd != got {
+		t.Fatalf("incorrect redirect, exp: %s, got: %s", exp, got)
 	}
 }
 
 func Test_FormRedirectHTTPS(t *testing.T) {
-	m := &MockStore{}
-	c := &mockClusterService{}
+	m := &MockStore{
+		leaderAddr: "foo:4002",
+	}
+	c := &mockClusterService{
+		apiAddr: "https://foo:4001",
+	}
+
 	s := New("127.0.0.1:0", m, c, nil)
 	req := mustNewHTTPRequest("http://qux:4001")
 
-	if rd := s.FormRedirect(req, "https://foo:4001"); rd != "https://foo:4001" {
-		t.Fatal("failed to form redirect for simple URL")
+	rd, err := s.FormRedirect(req)
+	if err != nil {
+		t.Fatalf("failed to form redirect: %s", err.Error())
+	}
+	if exp, got := "https://foo:4001", rd; exp != got {
+		t.Fatalf("incorrect redirect, exp: %s, got: %s", exp, got)
+	}
+}
+
+func Test_DoRedirect(t *testing.T) {
+	m := &MockStore{
+		leaderAddr: "foo:4002",
+	}
+	c := &mockClusterService{
+		apiAddr: "https://foo:4001",
+	}
+	s := New("127.0.0.1:0", m, c, nil)
+	req := mustNewHTTPRequest("http://qux:4001")
+
+	if s.DoRedirect(nil, req) {
+		t.Fatalf("incorrectly redirected")
+	}
+
+	req = mustNewHTTPRequest("http://qux:4001/db/query?redirect")
+	w := httptest.NewRecorder()
+	if !s.DoRedirect(w, req) {
+		t.Fatalf("incorrectly not redirected")
+	}
+	if exp, got := http.StatusMovedPermanently, w.Code; exp != got {
+		t.Fatalf("incorrect redirect code, exp: %d, got: %d", exp, got)
+	}
+	// check location header
+	if exp, got := "https://foo:4001/db/query?redirect", w.Header().Get("Location"); exp != got {
+		t.Fatalf("incorrect redirect location, exp: %s, got: %s", exp, got)
 	}
 }